backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# File handling
aiofiles==23.2.1
python-multipart==0.0.6

# Basic data processing (for Week1 setup)
pandas==2.1.3
numpy==1.25.2

# Configuration files
PyYAML==6.0.1

# AI/ML libraries for model execution
scvi-tools==0.20.3
scanpy==1.9.6
matplotlib==3.8.2
seaborn==0.13.0
torch==2.1.1

# Development tools
pytest==7.4.3 